把內存登入改成資料庫數據登入&密碼加密

+++ crowdfunding01-admin-parent/crowdfunding03-admin-component/src/main/java/com/dustin/crowd/mvc/handler/AssignHandler.java
@@ -31,7 +31,7 @@ public class AssignHandler {
 	private AuthService authService;
 
 	@RequestMapping("/assign/to/assign/role/page.html")
-	public String toAssignRolePage(@RequestParam("adminId") Integer adminId, ModelMap modelMap) {
+	public String toAssignRolePage(@RequestParam("adminId") Long adminId, ModelMap modelMap) {
 		// 1.查询已分配角色
 		List<Role> assignedRoleList = roleService.getAssignedRole(adminId);
